style : replace AWS code on service layer
