+++ src/config.ts
@@ -13,3 +13,4 @@ export const AWS_S3_REGION: string = config.AWS_S3_REGION;
 export const AWS_ACCESS_KEY_ID: string = config.AWS_ACCESS_KEY_ID;
 export const AWS_SECRET_ACCESS_KEY: string = config.AWS_SECRET_ACCESS_KEY;
 export const APP_DOMAIN: string = config.APP_DOMAIN;
+export const AWS_S3_BUCKET_NAME: string = config.AWS_S3_BUCKET_NAME;
