+++ src/services/image-service.ts
@@ -1,7 +1,14 @@
 import { Storage } from "@google-cloud/storage";
+import { S3Client, PutObjectCommand } from "@aws-sdk/client-s3";
 import path from "path";
 
 import { GOOGLE_CLOUD_BUCKET_NAME } from "../config";
+import {
+  AWS_S3_REGION,
+  AWS_ACCESS_KEY_ID,
+  AWS_SECRET_ACCESS_KEY,
+  AWS_S3_BUCKET_NAME,
+} from "../config";
 
 const storage = new Storage();
 
@@ -33,4 +40,28 @@ async function uploadGoogleCloud(
   return { url: fileUrl };
 }
 
-export default { uploadGoogleCloud };
+const s3Client = new S3Client({
+  region: AWS_S3_REGION,
+  credentials: {
+    accessKeyId: AWS_ACCESS_KEY_ID as string,
+    secretAccessKey: AWS_SECRET_ACCESS_KEY as string,
+  },
+});
+
+async function uploadAWS(file: Express.Multer.File): Promise<{ url: string }> {
+  const bucketName = AWS_S3_BUCKET_NAME;
+
+  const uploadParams = {
+    Bucket: bucketName,
+    Key: `${Date.now()}-${file.originalname}`,
+    Body: file.buffer,
+    ContentType: file.mimetype,
+  };
+
+  const result = await s3Client.send(new PutObjectCommand(uploadParams));
+  const uploadedImageUrl = `https://${bucketName}.s3.${process.env.AWS_REGION}.amazonaws.com/${uploadParams.Key}`;
+
+  return { url: uploadedImageUrl };
+}
+
+export default { uploadGoogleCloud, uploadAWS };
